+++ NEWS.rst
@@ -20,6 +20,7 @@ New Features
 * New contrib module `destructure` for Clojure-style destructuring.
 * Location of history file now configurable via environment variable `HY_HISTORY`
 * Added handling for "=" syntax in f-strings.
+* Repl init scripts with `HYSTARTUP` env var
 
 Bug Fixes
 ------------------------------